Make sensitive phone numbers not to be shown in call log history [1/2]
Considering sensitive phone numbers to be, hotlines dealing with
violence against women.
In the EU, these numbers do not show up in the phone bill.
In order to avoid these phone numbers to be listed in the Call Log, we
have implemented a new XML file which is parsed with phone numbers
from different countries to be filtered. This file needs to be copied to
device via desired Android.mk file in order to be able to process it.
The comparison is made checking the network MCC the SIM is connected to
in order to consider roaming and multisim scenarios.

Test: CallLogManagerTest.testDontLogCallsToSensitivePhoneNumber PASS

Change-Id: I4a59ff0577942ce56924f1a434ae0a3a38eacc62
Signed-off-by: Olaia Segovia <olaia.segovia@bq.com>